+++ src/frontend/custom_widgets/tracker_format_override.py
@@ -1,11 +1,11 @@
-from PySide6.QtCore import Signal, Slot, Qt
+from PySide6.QtCore import Qt, Signal, Slot
 from PySide6.QtWidgets import (
+    QCheckBox,
     QLabel,
+    QLineEdit,
     QSizePolicy,
     QVBoxLayout,
-    QLineEdit,
     QWidget,
-    QCheckBox,
 )
 
 from src.backend.token_replacer import ColonReplace
@@ -13,6 +13,7 @@ from src.frontend.custom_widgets.combo_box import CustomComboBox
 from src.frontend.custom_widgets.replacement_list_widget import (
     LoadedReplacementListWidget,
 )
+from src.frontend.utils import build_h_line
 
 
 class TrackerFormatOverride(QWidget):
@@ -71,11 +72,12 @@ class TrackerFormatOverride(QWidget):
         self.over_ride_inner_layout.setContentsMargins(6, 0, 0, 0)
         self.over_ride_inner_layout.addWidget(title_colon_replace_lbl)
         self.over_ride_inner_layout.addWidget(self.title_colon_replace)
+        self.over_ride_inner_layout.addWidget(build_h_line((6, 1, 6, 1)))
         self.over_ride_inner_layout.addWidget(over_ride_format_title_lbl)
         self.over_ride_inner_layout.addWidget(self.over_ride_format_title)
 
         self.over_rider_inner_nested_layout = QVBoxLayout()
-        self.over_rider_inner_nested_layout.setContentsMargins(6, 0, 0, 0)
+        self.over_rider_inner_nested_layout.setContentsMargins(20, 0, 0, 0)
         self.over_rider_inner_nested_layout.addWidget(
             over_ride_format_file_name_token_example_lbl
         )