+++ CHANGELOG.md
@@ -95,6 +95,7 @@
   - Text window is now read only.
 - Plugin Changes:
   - `pre_upload` plugin is no longer passed kwargs `media_file, mi_obj, source_path`. These can be gathered from the config object easily.
+- Improved the visuals of tracker format override widget.
 
 ### Fixed